Rework CI

+++ cpp/src/IceSSL/SChannelEngine.cpp
@@ -661,7 +661,7 @@ SChannel::SSLEngine::initialize()
             }
         }
 
-        for (int i = 0; i < certFiles.size(); ++i)
+        for (size_t i = 0; i < certFiles.size(); ++i)
         {
             string certFile = certFiles[i];
             string resolved;
